pyproject.toml
[tool.poetry]
name = "hangar-sdk"
version = "0.0.27"
description = "SDK for hangar cloud."
authors = ["Mihir Chintawar <mihir@tryhangar.com>","Kiran Illindala <kiran@tryhangar.com>"]
readme = "README.md"
license = "APACHE 2.0"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pydantic = "^2.6.2"
plumbum = "^1.8.2"
aioboto3 = "^12.3.0"
boto3 = "1.34.34"
cryptography = "^42.0.3"
paramiko = "^3.4.0"
mypy = "^1.8.0"
botocore-stubs = "^1.34.56"
boto3-stubs = "^1.34.55"
attrs = "^23.2.0"
pytest = "^8.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
types-python-dateutil = "^2.8.19.20240106"
ruff = "^0.3.0"
types-six = "^1.16.21.20240301"
types-setuptools = "^69.1.0.20240302"
types-aioboto3 = "^12.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
